make all implementations implement new GenericCache interface, remove typed packages
