+++ src/main/kotlin/io/github/kezhenxu94/cache/lite/impl/LRUCache.kt
@@ -16,28 +16,28 @@
 
 package io.github.kezhenxu94.cache.lite.impl
 
-import io.github.kezhenxu94.cache.lite.Cache
+import io.github.kezhenxu94.cache.lite.GenericCache
 
 /**
  * [LRUCache] flushes items that are **Least Recently Used** and keeps [minimalSize] items at most.
  */
-class LRUCache(private val delegate: Cache, private val minimalSize: Int = DEFAULT_SIZE) : Cache by delegate {
-  private val keyMap = object : LinkedHashMap<Any, Any>(minimalSize, .75f, true) {
-    override fun removeEldestEntry(eldest: MutableMap.MutableEntry<Any, Any>): Boolean {
+class LRUCache<K, V>(private val delegate: GenericCache<K, V>, private val minimalSize: Int = DEFAULT_SIZE) : GenericCache<K, V> by delegate {
+  private val keyMap = object : LinkedHashMap<K, Boolean>(minimalSize, .75f, true) {
+    override fun removeEldestEntry(eldest: MutableMap.MutableEntry<K, Boolean>): Boolean {
       val tooManyCachedItems = size > minimalSize
       if (tooManyCachedItems) eldestKeyToRemove = eldest.key
       return tooManyCachedItems
     }
   }
 
-  private var eldestKeyToRemove: Any? = null
+  private var eldestKeyToRemove: K? = null
 
-  override fun set(key: Any, value: Any) {
+  override fun set(key: K, value: V) {
     delegate[key] = value
     cycleKeyMap(key)
   }
 
-  override fun get(key: Any): Any? {
+  override fun get(key: K): V? {
     keyMap[key]
     return delegate[key]
   }
@@ -47,7 +47,7 @@ class LRUCache(private val delegate: Cache, private val minimalSize: Int = DEFAU
     delegate.clear()
   }
 
-  private fun cycleKeyMap(key: Any) {
+  private fun cycleKeyMap(key: K) {
     keyMap[key] = PRESENT
     eldestKeyToRemove?.let { delegate.remove(it) }
     eldestKeyToRemove = null
