+++ src/main/kotlin/io/github/kezhenxu94/cache/lite/impl/SoftCache.kt
@@ -16,7 +16,7 @@
 
 package io.github.kezhenxu94.cache.lite.impl
 
-import io.github.kezhenxu94.cache.lite.Cache
+import io.github.kezhenxu94.cache.lite.GenericCache
 
 import java.lang.ref.ReferenceQueue
 import java.lang.ref.SoftReference
@@ -24,7 +24,7 @@ import java.lang.ref.SoftReference
 /**
  * [SoftCache] caches items with a [SoftReference] wrapper.
  */
-class SoftCache(private val delegate: Cache) : Cache by delegate {
+class SoftCache(private val delegate: GenericCache<Any, Any>) : GenericCache<Any, Any> by delegate {
   private val referenceQueue = ReferenceQueue<Any>()
 
   private class SoftEntry internal constructor(
